преобразование в репозиторий бд, переименование методов репозитория в сервисе

+++ src/main/java/home/vs/testtask/service/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package home.vs.testtask.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,23 +19,32 @@ public class UserServiceImpl implements UserService {
         this.repository = repository;
     }
 
+    @Override
     public User getByLogin(String login) {
-        return this.repository.getByLogin(login);
+        return this.repository.findByLogin(login).orElse(null);
     }
 
+    @Override
     public User getById(Long id) {
         return this.repository.getById(id);
     }
 
+    @Override
     public List<User> getAll() {
-        return this.repository.getAll();
+        return this.repository.findAll();
     }
 
+    @Override
     public User create(User user) {
-        return repository.create(user);
+        return repository.saveAndFlush(user);
     }
 
+    @Override
     public User deleteById(Long id) {
-        return this.repository.deleteById(id);
+        Optional<User> user = this.repository.findById(id);
+        if (!user.isEmpty()) {
+            this.repository.deleteById(id);
+        }
+        return null;
     }
 }
